SPR-7.3: Fix bug strength of controlnet is float type not int type

+++ src/core/module/generate-image/GenerateImageController.ts
@@ -43,7 +43,7 @@ export class GenerateImageController {
           controlNetType: control_net.controlNetType,
           image: data_images.controlNetImages[_index].buffer,
           strength: control_net.strength
-            ? parseInt(control_net.strength.toString())
+            ? parseFloat(control_net.strength.toString())
             : CONTROL_NET_DEFAULT_STRENGTH,
           isPreprocessor: control_net.isPreprocessor
             ? control_net.isPreprocessor.toString() === 'true'
@@ -79,7 +79,7 @@ export class GenerateImageController {
           controlNetType: control_net.controlNetType,
           image: data_images.controlNetImages[_index].buffer,
           strength: control_net.strength
-            ? parseInt(control_net.strength.toString())
+            ? parseFloat(control_net.strength.toString())
             : CONTROL_NET_DEFAULT_STRENGTH,
           isPreprocessor: control_net.isPreprocessor
             ? control_net.isPreprocessor.toString() === 'true'
